backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-extra-fields==3.0.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
